requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9  # PostgreSQL adapter
types-requests>=2.31.0  # Type stubs for requests
mypy>=1.8.0  # Optional: for type checking
